Removed batch method and made default record method work as batch, added optional force push parameter.
